86cymwjqt - Adjust shadcn folder structure and merge agent, landlord and tenant side nav bar link definitions
