+++ app/client/ui-modules/guest-landing-page/GuestLandingPage.tsx
@@ -6,10 +6,9 @@ import { selectGuestLandingPageUiState } from "./state/reducers/guest-landing-pa
 import { useSelector } from "react-redux";
 import { SideNavBar } from "../navigation-bars/side-nav-bars/SideNavbar";
 import { TopNavbar } from "../navigation-bars/TopNavbar";
-
-import { Button } from "../theming/components/shadcn/Button";
-import { Input } from "../theming/components/shadcn/Input";
-import { agentLinks } from "../navigation-bars/side-nav-bars/AgentSideNav";
+import { Button } from "../theming-shadcn/Button";
+import { Input } from "../theming-shadcn/Input";
+import { agentLinks } from "../navigation-bars/side-nav-bars/side-nav-link-definitions";
 export function GuestLandingPage(): React.JSX.Element {
   const GuestLandingPageUiState: GuestLandingPageUiState = useSelector(
     selectGuestLandingPageUiState
