+++ app/client/ui-modules/navigation-bars/BottonNavbar.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router";
 import { PropManagerLogoIcon } from "../theming/components/logo/PropManagerLogoIcon";
 import { PropManagerLogoText } from "../theming/components/logo/PropManagerLogoText";
-import { NavBarLink } from "./side-nav-bars/NavBarLink";
+import { NavBarLink } from "./side-nav-bars/components/NavBarLink";
 
 export function BottomNavbar(): React.JSX.Element {
   return (
