+++ app/client/ui-modules/navigation-bars/TopNavbar.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Link } from "react-router";
 import { PropManagerLogoIcon } from "../theming/components/logo/PropManagerLogoIcon";
 import { PropManagerLogoText } from "../theming/components/logo/PropManagerLogoText";
-import { Button } from "../theming/components/shadcn/Button";
+import { Button } from "../theming-shadcn/Button";
 interface TopNavbarProps {
   onSideBarOpened: React.Dispatch<React.SetStateAction<boolean>>;
 }
