--- app/client/ui-modules/navigation-bars/side-nav-bars/TenantSideNav.tsx
@@ -1,22 +1,0 @@
-import React from "react";
-import { MultipleHousesIcon } from "../../theming/icons/MultipleHousesIcon";
-import { CalendarIcon } from "../../theming/icons/CalendarIcon";
-
-export const tenantLinks = [
-  { to: "/", label: "Overview" },
-  {
-    to: "/properties",
-    label: "Managed Properties",
-    icon: <MultipleHousesIcon variant="light" />,
-  },
-  {
-    to: "/calendar",
-    label: "Calendar",
-    icon: <CalendarIcon variant="light" />,
-  },
-  {
-    to: "/tasks",
-    label: "Tasks",
-    icon: <CalendarIcon variant="light" />,
-  },
-];
